started refactoring tests
